refactor: split large files to meet 400-line limit
- Extract terminal_commands.py from terminal.py (commands + themes)
- Extract checkpoints.py from stage.py (checkpoint detection)
- Extract audio.py from runner.py (FFmpeg operations)
- Refactor preview.py with helper functions
- Update CI to use strict function length thresholds (warn 8, error 12)

File sizes after split:
- terminal.py: 466 → 334 lines
- stage.py: 312 → 208 lines
- runner.py: 333 → 189 lines
