+++ src/demorec/audio.py
@@ -1,0 +1,155 @@
+"""Audio and video processing utilities using FFmpeg.
+
+Provides functions for mixing, concatenating, and processing audio/video files.
+"""
+
+import json
+import shutil
+import subprocess
+from pathlib import Path
+
+
+def run_ffmpeg(cmd: list[str], error_msg: str):
+    """Run an FFmpeg command, raising RuntimeError on failure."""
+    result = subprocess.run(cmd, capture_output=True, text=True)
+    if result.returncode != 0:
+        raise RuntimeError(f"{error_msg}: {result.stderr}")
+
+
+def write_concat_file(file_path: Path, files: list[Path]):
+    """Write an FFmpeg concat file list."""
+    with open(file_path, "w") as f:
+        for item in files:
+            f.write(f"file '{item}'\n")
+
+
+def get_duration(media_path: Path) -> float:
+    """Get duration of a media file in seconds."""
+    cmd = [
+        "ffprobe", "-v", "quiet", "-print_format", "json",
+        "-show_format", str(media_path),
+    ]
+    result = subprocess.run(cmd, capture_output=True, text=True)
+    if result.returncode == 0:
+        data = json.loads(result.stdout)
+        return float(data.get("format", {}).get("duration", 0))
+    return 0.0
+
+
+def concat_audio_files(audio_files: list[Path], output: Path, temp_dir: Path) -> Path:
+    """Concatenate multiple audio files into one."""
+    concat_file = temp_dir / "narration_concat.txt"
+    write_concat_file(concat_file, audio_files)
+
+    cmd = [
+        "ffmpeg", "-y", "-f", "concat", "-safe", "0",
+        "-i", str(concat_file), "-c", "copy", str(output),
+    ]
+    run_ffmpeg(cmd, "Audio concat failed")
+    return output
+
+
+def overlay_audio(video: Path, audio: Path, output: Path):
+    """Overlay audio track onto video."""
+    cmd = [
+        "ffmpeg", "-y",
+        "-i", str(video), "-i", str(audio),
+        "-c:v", "copy", "-c:a", "aac", "-shortest", str(output),
+    ]
+    run_ffmpeg(cmd, "Audio overlay failed")
+
+
+def mix_audio_timed(video_path: Path, narrations: list, output: Path):
+    """Mix narration audio with video at correct timestamps.
+
+    Args:
+        video_path: Input video file
+        narrations: List of objects with audio_path, start_time, duration
+        output: Output video file path
+    """
+    if not narrations:
+        shutil.copy(video_path, output)
+        return
+
+    video_duration = get_duration(video_path)
+    filter_complex = _build_audio_filter(narrations)
+
+    inputs = ["-i", str(video_path)]
+    for n in narrations:
+        inputs.extend(["-i", str(n.audio_path)])
+
+    cmd = [
+        "ffmpeg", "-y", *inputs,
+        "-filter_complex", filter_complex,
+        "-map", "0:v", "-map", "[aout]",
+        "-c:v", "copy", "-c:a", "aac",
+        "-t", str(video_duration), str(output),
+    ]
+    run_ffmpeg(cmd, "Audio mixing failed")
+
+
+def _build_audio_filter(narrations: list) -> str:
+    """Build FFmpeg filter_complex for mixing narrations."""
+    filter_parts = []
+
+    for i, n in enumerate(narrations):
+        delay_ms = int(max(0, n.start_time) * 1000)
+        filter_parts.append(f"[{i + 1}:a]adelay={delay_ms}|{delay_ms}[a{i}]")
+
+    if len(narrations) == 1:
+        return f"{filter_parts[0]}; [a0]apad[aout]"
+
+    mix_inputs = "".join(f"[a{i}]" for i in range(len(narrations)))
+    amix_opts = "duration=longest:normalize=0:dropout_transition=0"
+    amix_filter = f"{mix_inputs}amix=inputs={len(narrations)}:{amix_opts}[aout]"
+    return "; ".join(filter_parts) + f"; {amix_filter}"
+
+
+def format_srt_time(seconds: float) -> str:
+    """Format time in SRT format (HH:MM:SS,mmm)."""
+    hours = int(seconds // 3600)
+    minutes = int((seconds % 3600) // 60)
+    secs = int(seconds % 60)
+    millis = int((seconds % 1) * 1000)
+    return f"{hours:02d}:{minutes:02d}:{secs:02d},{millis:03d}"
+
+
+def split_caption(text: str, max_len: int = 42) -> list[str]:
+    """Split caption text into lines of max_len characters."""
+    if len(text) <= max_len:
+        return [text]
+
+    lines = []
+    current_line = ""
+
+    for word in text.split():
+        if len(current_line) + len(word) + 1 <= max_len:
+            current_line = f"{current_line} {word}".strip()
+        else:
+            if current_line:
+                lines.append(current_line)
+            current_line = word
+
+    if current_line:
+        lines.append(current_line)
+
+    return lines
+
+
+def generate_srt(narrations: list, output_path: Path):
+    """Generate SRT subtitle file from timed narrations.
+
+    Args:
+        narrations: List of objects with text, start_time, duration
+        output_path: Output SRT file path
+    """
+    with open(output_path, "w", encoding="utf-8") as f:
+        for i, n in enumerate(narrations, 1):
+            start = max(0, n.start_time)
+            end = start + n.duration
+            lines = split_caption(n.text)
+            caption_text = "\n".join(lines)
+
+            f.write(f"{i}\n")
+            f.write(f"{format_srt_time(start)} --> {format_srt_time(end)}\n")
+            f.write(f"{caption_text}\n\n")
